requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.4
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
requests==2.33.1
requests-oauthlib==2.0.0
sqlparse==0.5.5
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
resend>=2.0.0
